+++ internal/settings/push_repo.go
@@ -3,20 +3,28 @@ package settings
 import (
 	"errors"
 	"fmt"
-	"sync"
+	"strings"
 
 	vaultapi "github.com/hashicorp/vault/api"
+	"github.com/rs/zerolog/log"
 )
 
 const (
 	keyData  = "data"
+	keysData = "keys"
 	keyToken = "token"
 )
 
 type vaultReadWriter interface {
+	List(path string) (*vaultapi.Secret, error)
 	Read(path string) (*vaultapi.Secret, error)
 	Write(path string, data map[string]interface{}) (*vaultapi.Secret, error)
 	Delete(path string) (*vaultapi.Secret, error)
+}
+
+type PushDetails struct {
+	DeviceUUID string
+	Token      string
 }
 
 var (
@@ -25,36 +33,85 @@ var (
 )
 
 // PushRepo provides capability to store and get keys in Vault.
-// Also this struct has in-memory cache to reduce database queries
 type PushRepo struct {
 	cli      vaultReadWriter
 	basePath string
-
-	cache map[string]string
-	mux   sync.Mutex
 }
 
 func NewPushRepo(cli vaultReadWriter, path string) *PushRepo {
 	return &PushRepo{
 		cli:      cli,
-		basePath: path,
-		cache:    make(map[string]string),
+		basePath: strings.TrimRight(path, "/"),
 	}
 }
 
-func (s *PushRepo) getPath(userID string) string {
-	return fmt.Sprintf("%s%s", s.basePath, userID)
+func (s *PushRepo) getPathV1(userID string) string {
+	return fmt.Sprintf("%s/%s", s.basePath, userID)
 }
 
+func (s *PushRepo) getPathByUser(userID string) string {
+	return fmt.Sprintf("%s/tokens_by_devices/%s", s.basePath, userID)
+}
+
+func (s *PushRepo) getPathByUserDevice(userID, deviceUUID string) string {
+	return fmt.Sprintf("%s/%s", s.getPathByUser(userID), deviceUUID)
+}
+
+func (s *PushRepo) GetListByUserID(userID string) ([]PushDetails, error) {
+	sec, err := s.cli.List(s.getPathByUser(userID))
+	if err != nil {
+		return nil, err
+	}
+
+	if sec == nil {
+		// let's try fallback logic
+		token, err := s.GetByUserID(userID)
+		if err != nil && !errors.Is(err, ErrTokenNotFound) {
+			return nil, fmt.Errorf("fallback by user: %w", err)
+		}
+
+		if errors.Is(err, ErrTokenNotFound) {
+			return nil, nil
+		}
+
+		return []PushDetails{
+			{
+				DeviceUUID: "default_device",
+				Token:      token,
+			},
+		}, nil
+	}
+
+	data, ok := sec.Data[keysData].([]interface{})
+	if !ok {
+		return nil, ErrUnableToCastData
+	}
+
+	result := make([]PushDetails, 0, len(data))
+	for idx := range data {
+		deviceUUID, ok := data[idx].(string)
+		if !ok {
+			return nil, fmt.Errorf("cast device uuid: %w", ErrUnableToCastData)
+		}
+
+		token, err := s.GetByUserAndDevice(userID, deviceUUID)
+		if err != nil {
+			return nil, fmt.Errorf("get token by device : %w", err)
+		}
+
+		result = append(result, PushDetails{
+			DeviceUUID: deviceUUID,
+			Token:      token,
+		})
+	}
+
+	return result, nil
+}
+
+// GetByUserID returns saved user token
+// @deprecated: use GetByUserAndDevice instead
 func (s *PushRepo) GetByUserID(userID string) (string, error) {
-	s.mux.Lock()
-	defer s.mux.Unlock()
-
-	if token, ok := s.cache[userID]; ok {
-		return token, nil
-	}
-
-	sec, err := s.cli.Read(s.getPath(userID))
+	sec, err := s.cli.Read(s.getPathV1(userID))
 	if err != nil {
 		return "", err
 	}
@@ -73,16 +130,51 @@ func (s *PushRepo) GetByUserID(userID string) (string, error) {
 		return "", ErrUnableToCastData
 	}
 
-	s.cache[userID] = token
+	return token, nil
+}
+
+func (s *PushRepo) GetByUserAndDevice(userID, deviceUUID string) (token string, err error) {
+	sec, err := s.cli.Read(s.getPathByUserDevice(userID, deviceUUID))
+	if err != nil {
+		return "", err
+	}
+
+	defer func() {
+		if err == nil {
+			return
+		}
+
+		// fallback for getting by old version and resaving by new path
+		token, err = s.GetByUserID(userID)
+		if err != nil {
+			return
+		}
+
+		if errSave := s.Save(userID, deviceUUID, token); errSave != nil {
+			log.Err(err).Msg("failed to resave token")
+		}
+	}()
+
+	if sec == nil {
+		return "", ErrTokenNotFound
+	}
+
+	data, ok := sec.Data[keyData].(map[string]interface{})
+	if !ok {
+		return "", ErrUnableToCastData
+	}
+
+	token, ok = data[keyToken].(string)
+	if !ok {
+		return "", ErrUnableToCastData
+	}
 
 	return token, nil
 }
 
-func (s *PushRepo) Save(userID, token string) error {
-	s.mux.Lock()
-	defer s.mux.Unlock()
-
-	_, err := s.cli.Write(s.getPath(userID), map[string]interface{}{
+// Save storing provided token for user by device
+func (s *PushRepo) Save(userID, deviceUUID, token string) error {
+	_, err := s.cli.Write(s.getPathByUserDevice(userID, deviceUUID), map[string]interface{}{
 		keyData: map[string]interface{}{
 			keyToken: token,
 		},
@@ -91,21 +183,15 @@ func (s *PushRepo) Save(userID, token string) error {
 		return err
 	}
 
-	s.cache[userID] = token
-
 	return nil
 }
 
-func (s *PushRepo) Delete(userID string) error {
-	s.mux.Lock()
-	defer s.mux.Unlock()
-
-	_, err := s.cli.Delete(s.getPath(userID))
+// Delete remove token from storage by user and device
+func (s *PushRepo) Delete(userID, deviceUUID string) error {
+	_, err := s.cli.Delete(s.getPathByUserDevice(userID, deviceUUID))
 	if err != nil {
 		return err
 	}
-
-	delete(s.cache, userID)
 
 	return nil
 }