+++ CHANGELOG.md
@@ -7,6 +7,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Separate storing tokens by device
+
+### Added
+- Endpoint for getting list of tokens
+
 ## [0.0.32] - 2024-05-13
 
 ### Fixed